+++ 3D-Project/ConstantBuffer.h
@@ -7,7 +7,7 @@ class ConstantBuffer
 {
 private:
 	ID3D11Buffer* m_buffer;
-	size_t m_size;
+	UINT m_size;
 
 public:
 	ConstantBuffer() = default;
@@ -17,7 +17,7 @@ public:
 	bool Init(ID3D11Device* device, size_t byteSize, void* initData = nullptr);
 
 	ID3D11Buffer* GetBuffer() const;
-	size_t GetSize() const;
+	UINT GetSize() const;
 
 	void Update(ID3D11DeviceContext* context, void* data);
 };
